+++ api/src/main/java/com/discoodle/api/service/MessagesService.java
@@ -22,22 +22,27 @@ public class MessagesService {
    }
 
    public void pinMessage(Long message_id) {
-      messageRepository.pinMessage(message_id);
+      if(messageRepository.existsById(message_id))
+         messageRepository.pinMessage(message_id);
    }
 
    public void unpinMessage(Long message_id) {
-      messageRepository.unpinMessage(message_id);
+      if(messageRepository.existsById(message_id))
+         messageRepository.unpinMessage(message_id);
    }
 
    public void deleteMessage(Long message_id) {
-      messageRepository.deleteById(message_id);
+      if(messageRepository.existsById(message_id))
+         messageRepository.deleteById(message_id);
    }
 
    public void editMessage(Long message_id, String content) {
-      messageRepository.editMessage(message_id, content);
+      if(messageRepository.existsById(message_id))
+         messageRepository.editMessage(message_id, content);
    }
 
    public void setEdited(Long message_id) {
-      messageRepository.setEdited(message_id);
+      if(messageRepository.existsById(message_id))
+         messageRepository.setEdited(message_id);
    }
 }
